// src/api/generated/models/proficencyTestModel.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * Tg-Api 
 * OpenAPI spec version: 1.0.0
 */
import type { DevProfileModel } from './devProfileModel';

export interface ProficencyTestModel {
  id: string;
  dev_profile_id: string;
  dev_profile?: DevProfileModel;
  seniority_level: string;
  specialty: string;
  /** @nullable */
  backend_category?: string | null;
  /** @nullable */
  frontend_category?: string | null;
  status: string;
  /** @nullable */
  score?: number | null;
  /** @nullable */
  max_score?: number | null;
  solicitation_date: string;
  created_at: string;
  updated_at: string;
}